:recycle: refactor(`<DocumentsList>`): Parameterize fetching hook.
