+++ imports/ui/documents/DocumentsList.ts
@@ -1,0 +1,6 @@
+import useDocuments from './useDocuments';
+import makeDocumentsList from './makeDocumentsList';
+
+const DocumentsList = makeDocumentsList(useDocuments);
+
+export default DocumentsList;
